pyproject.toml
[project]
name = "wealthsimple-deepsearch-extension"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "gradio>=6.6.0",
    "python-dotenv>=1.2.1",
    "smolagents[litellm,toolkit]>=1.24.0",
    "uvicorn[standard]>=0.32.0",
    "yfinance>=1.2.0",
]
